+++ tests/conftest.py
@@ -5,7 +5,7 @@ from monty.serialization import loadfn
 from smol.cofe import ClusterSubspace
 from smol.moca import CanonicalEnsemble, MuSemiGrandEnsemble, \
     FuSemiGrandEnsemble
-from smol.moca.processor import CEProcessor
+from smol.moca.processor import ClusterExpansionProcessor
 
 # load test data files and set them up as fixtures
 DATA_DIR = os.path.join(os.path.dirname(__file__), 'data')
@@ -39,7 +39,7 @@ def single_subspace():
 @pytest.fixture(params=ensembles)
 def ensemble(cluster_subspace, request):
     coefs = np.random.random(cluster_subspace.num_corr_functions)
-    proc = CEProcessor(cluster_subspace, 4 * np.eye(3), coefs)
+    proc = ClusterExpansionProcessor(cluster_subspace, 4 * np.eye(3), coefs)
     if request.param is MuSemiGrandEnsemble:
         kwargs = {'chemical_potentials':
                   {sp: 0.3 for space in proc.unique_site_spaces
@@ -52,5 +52,5 @@ def ensemble(cluster_subspace, request):
 @pytest.fixture
 def single_canonical_ensemble(single_subspace):
     coefs = np.random.random(single_subspace.num_corr_functions)
-    proc = CEProcessor(single_subspace, 4 * np.eye(3), coefs)
+    proc = ClusterExpansionProcessor(single_subspace, 4 * np.eye(3), coefs)
     return CanonicalEnsemble(proc)